+++ core/src/elements/ons-splitter/index.js
@@ -20,6 +20,7 @@ import ModifierUtil from '../../ons/internal/modifier-util';
 import AnimatorFactory from '../../ons/internal/animator-factory';
 import OverlaySplitterAnimator from './overlay-animator';
 import PushSplitterAnimator from './push-animator';
+import RevealSplitterAnimator from './reveal-animator';
 import BaseElement from '../../ons/base-element';
 import deviceBackButtonDispatcher from '../../ons/device-back-button-dispatcher';
 import contentReady from '../../ons/content-ready';
@@ -27,7 +28,8 @@ import contentReady from '../../ons/content-ready';
 const _animatorDict = {
   default: OverlaySplitterAnimator,
   overlay: OverlaySplitterAnimator,
-  push: PushSplitterAnimator
+  push: PushSplitterAnimator,
+  reveal: RevealSplitterAnimator
 };
 
 /**